Stop providing ruby-ruby2-keywords for now
ruby-ruby2-keywords is its own packages, and is now included in Ruby
itself. If libruby3.1 Provides: ruby-ruby2-keywords, the actual binary
will not be installed, and builds and tests will fail on ruby3.0 after
ruby3.1 is added as a supported version: build dependency resolution
will assume ruby-ruby2-keywords is provided by libruby3.1, and ruby3.0
will not actually have access to it.

This can be reverted after ruby3.1 becomes the default Ruby and support
for ruby3.0 is removed.

Gbp-Dch: full