Fix CVE-2025-27219
In the CGI gem, the CGI::Cookie.parse method in the CGI library contains
a potential Denial of Service (DoS) vulnerability.  The method does not
impose any limit on the length of the raw cookie value it processes.
This oversight can lead to excessive resource consumption when parsing
extremely large cookies.

  - d/p/CVE-2025-27219.patch

Gbp-Dch: Full