d/control: make libruby3.3 depend on versioned ruby-{csv,ruby2-keywords}
Those 2 gems used to have the same version in libruby3.1 and in their
own source packages, and when a user tried to upgrade from bookworm to
trixie the libruby3.1 was kept because it would satisfy the depedencies
without installing a new package.

Adding them with a version constraint to avoid keeping libruby3.1 around
after the upgrade to ruby3.3.

Closes: #1099067

Gbp-Dch: Full