libruby3.3: drop dependencies on ruby-test-unit and ruby-minitest
This helps upgrades from bookworm, given that those used to be provided
by libruby3.1 but are no longer provided by libruby3.3, and those
dependencies cause libruby3.1* to be kept around on upgrades. These
packages are only needed for development, and for building packages
gem2deb-test-runner already depend on both of those anyway.